requirement.txt

google-cloud-bigquery==3.10.0
functions-framework==3.3.0